# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: paper-trader
    description: Paper trading dashboard and simulation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Legacy BTC-only endpoints (backward compat) ────────────────────────────
  /paper-trader/state:
    get:
      operationId: getPaperTraderState
      tags: [paper-trader]
      summary: Get current BTC paper trading state
      responses:
        "200":
          description: Current BTC paper trading state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTraderState"
  /paper-trader/refresh:
    post:
      operationId: refreshPaperTrader
      tags: [paper-trader]
      summary: Refresh BTC market data and evaluate strategy
      responses:
        "200":
          description: Refreshed BTC paper trading state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTraderState"
        "502":
          description: Kraken public API unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /paper-trader/trades:
    get:
      operationId: listPaperTrades
      tags: [paper-trader]
      summary: List simulated BTC trades
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Simulated BTC trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperTrade"
  /paper-trader/reset:
    post:
      operationId: resetPaperTrader
      tags: [paper-trader]
      summary: Reset the BTC virtual account
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPaperTraderInput"
      responses:
        "200":
          description: Reset BTC paper trading state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTraderState"

  # ── Multi-coin endpoints ───────────────────────────────────────────────────
  /paper-trader/multi-state:
    get:
      operationId: getMultiCoinState
      tags: [paper-trader]
      summary: Get paper trading state for all six instruments
      description: Returns BTC, ETH, SOL, XRP plus GOLD and SILVER (paper trading, unvalidated strategy) states in one call.
      responses:
        "200":
          description: All coin states
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiCoinState"
  /paper-trader/multi-refresh:
    post:
      operationId: refreshMultiCoin
      tags: [paper-trader]
      summary: Refresh market data and strategy for all six instruments
      description: Fetches Kraken public data for BTC, ETH, SOL, and XRP, plus gold-api.com spot prices and Yahoo Finance futures candles for GOLD and SILVER (paper trading, unvalidated strategy), and evaluates each strategy.
      responses:
        "200":
          description: Refreshed multi-coin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiCoinState"
        "502":
          description: Kraken public API unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /paper-trader/engine:
    get:
      operationId: getEngineStatus
      tags: [paper-trader]
      summary: Background strategy-scan engine status (server-side scheduler)
      responses:
        "200":
          description: Engine status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineStatus"

  /paper-trader/portfolio:
    get:
      operationId: getPortfolioSummary
      tags: [paper-trader]
      summary: Get combined portfolio summary across all six instruments
      responses:
        "200":
          description: Portfolio summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSummary"
  /paper-trader/activity:
    get:
      operationId: listActivityLog
      tags: [paper-trader]
      summary: Get the recent activity log
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Activity log entries newest-first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
  /paper-trader/all-trades:
    get:
      operationId: listAllTrades
      tags: [paper-trader]
      summary: List simulated trades for all coins
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 500
            default: 200
      responses:
        "200":
          description: All simulated trades newest-first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperTrade"
  /paper-trader/reset-all:
    post:
      operationId: resetAllCoins
      tags: [paper-trader]
      summary: Reset all six virtual accounts
      description: Resets BTC, ETH, SOL, and XRP accounts to the requested starting balance and clears history. GOLD and SILVER always reset to their fixed $100 baseline.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPaperTraderInput"
      responses:
        "200":
          description: All coins reset — returns new multi-coin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiCoinState"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Indicators ─────────────────────────────────────────────────────────
    IndicatorSnapshot:
      type: object
      properties:
        rsi:
          type: ["number", "null"]
        macd:
          type: ["number", "null"]
        macdSignal:
          type: ["number", "null"]
        atr:
          type: ["number", "null"]
        ema20:
          type: ["number", "null"]
        ema50:
          type: ["number", "null"]
        volume:
          type: ["number", "null"]
      required: [rsi, macd, macdSignal, atr, ema20, ema50, volume]

    # ── Strategy conditions breakdown ──────────────────────────────────────
    StrategyCondition:
      type: object
      properties:
        name:
          type: string
        currentValue:
          type: string
        requiredValue:
          type: string
        pass:
          type: boolean
      required: [name, currentValue, requiredValue, pass]
    StrategyConditions:
      type: object
      properties:
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/StrategyCondition"
        passCount:
          type: number
        totalCount:
          type: number
        bias:
          type: string
          enum: [LONG, SHORT, NEUTRAL]
        signal:
          type: string
          enum: [LONG, SHORT, NO_TRADE]
        oneHourTrend:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        fourHourTrend:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        indicators:
          $ref: "#/components/schemas/IndicatorSnapshot"
      required: [conditions, passCount, totalCount, bias, signal, oneHourTrend, fourHourTrend, indicators]

    # ── Proposed trade (when signal is present but no position open) ────────
    ProposedTrade:
      type: object
      properties:
        direction:
          type: string
          enum: [LONG, SHORT]
        entry:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        riskAmount:
          type: number
        rewardAmount:
          type: number
        rrRatio:
          type: number
        quantity:
          type: number
      required: [direction, entry, stopLoss, takeProfit, riskAmount, rewardAmount, rrRatio, quantity]

    # ── Open position ───────────────────────────────────────────────────────
    OpenPosition:
      type: object
      properties:
        direction:
          type: string
          enum: [LONG, SHORT]
        entry:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        quantity:
          type: number
        riskAmount:
          type: number
        openedAt:
          type: string
        entryRsi:
          type: ["number", "null"]
        entryMacd:
          type: ["number", "null"]
        entryAtr:
          type: ["number", "null"]
        trend4h:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        trend1h:
          type: ["string", "null"]
        longScore:
          type: ["number", "null"]
        shortScore:
          type: ["number", "null"]
        entryThreshold:
          type: ["number", "null"]
        unrealisedPnl:
          type: ["number", "null"]
        unrealisedPct:
          type: ["number", "null"]
        currentPrice:
          type: ["number", "null"]
      required: [direction, entry, stopLoss, takeProfit, quantity, riskAmount, openedAt, entryRsi, entryMacd, entryAtr, trend4h]

    # ── Trades ──────────────────────────────────────────────────────────────
    PaperTrade:
      type: object
      properties:
        id:
          type: number
        coin:
          type: string
        openedAt:
          type: string
        closedAt:
          type: string
        direction:
          type: string
          enum: [LONG, SHORT]
        entry:
          type: number
        exit:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        rsi:
          type: ["number", "null"]
        macd:
          type: ["number", "null"]
        atr:
          type: ["number", "null"]
        trend4h:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        profitLoss:
          type: number
        accountBalance:
          type: number
        exitReason:
          type: string
          enum: [TAKE_PROFIT, STOP_LOSS, SIGNAL_REVERSAL, MANUAL]
        riskAmount:
          type: ["number", "null"]
        rMultiple:
          type: ["number", "null"]
        pnlPct:
          type: ["number", "null"]
        durationSeconds:
          type: ["number", "null"]
        result:
          type: ["string", "null"]
          enum: [WIN, LOSS, BREAKEVEN, null]
        entryScore:
          type: ["number", "null"]
        passCount:
          type: ["number", "null"]
        trend1h:
          type: ["string", "null"]
        entryMode:
          type: ["string", "null"]
        entryConditions:
          type: ["string", "null"]
        longScore:
          type: ["number", "null"]
        shortScore:
          type: ["number", "null"]
        entryThreshold:
          type: ["number", "null"]
      required: [id, coin, openedAt, closedAt, direction, entry, exit, stopLoss, takeProfit, rsi, macd, atr, trend4h, profitLoss, accountBalance, exitReason]

    # ── Metrics ─────────────────────────────────────────────────────────────
    PaperTraderMetrics:
      type: object
      properties:
        virtualBalance:
          type: number
        startingBalance:
          type: number
        totalProfitLoss:
          type: number
        roi:
          type: number
        numberOfTrades:
          type: number
        wins:
          type: number
        losses:
          type: number
        winRate:
          type: number
        profitFactor:
          type: number
        maximumDrawdown:
          type: number
        dailyLoss:
          type: number
        consecutiveLosses:
          type: number
      required: [virtualBalance, startingBalance, totalProfitLoss, roi, numberOfTrades, wins, losses, winRate, profitFactor, maximumDrawdown, dailyLoss, consecutiveLosses]

    # ── Single-coin state ───────────────────────────────────────────────────
    PaperTraderState:
      type: object
      properties:
        coin:
          type: string
        market:
          type: object
          properties:
            pair:
              type: string
            currentPrice:
              type: ["number", "null"]
            updatedAt:
              type: string
            lastCompletedCandleAt:
              type: ["string", "null"]
          required: [pair, currentPrice, updatedAt, lastCompletedCandleAt]
        signal:
          type: string
          enum: [LONG, SHORT, NO_TRADE]
        oneHourTrend:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        fourHourTrend:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        indicators:
          $ref: "#/components/schemas/IndicatorSnapshot"
        strategyConditions:
          oneOf:
            - $ref: "#/components/schemas/StrategyConditions"
            - type: "null"
        proposedTrade:
          oneOf:
            - $ref: "#/components/schemas/ProposedTrade"
            - type: "null"
        directional:
          oneOf:
            - $ref: "#/components/schemas/DirectionalEvaluation"
            - type: "null"
        executionDiagnostics:
          oneOf:
            - $ref: "#/components/schemas/ExecutionDiagnostics"
            - type: "null"
        opportunity:
          $ref: "#/components/schemas/Opportunity"
        position:
          oneOf:
            - $ref: "#/components/schemas/OpenPosition"
            - type: "null"
        metrics:
          $ref: "#/components/schemas/PaperTraderMetrics"
        risk:
          type: object
          properties:
            dailyLossLimit:
              type: number
            maximumConsecutiveLosses:
              type: number
            riskPerTrade:
              type: number
            rewardToRisk:
              type: number
            pollingSeconds:
              type: number
          required: [dailyLossLimit, maximumConsecutiveLosses, riskPerTrade, rewardToRisk, pollingSeconds]
        recentTrades:
          type: array
          items:
            $ref: "#/components/schemas/PaperTrade"
        botStatus:
          type: string
          enum: [READY, WAITING_FOR_DATA, RISK_PAUSED, API_ERROR, MONITORING]
        message:
          type: string
        scanNote:
          oneOf:
            - type: string
            - type: "null"
          description: Set when scan candle data (Yahoo Finance) is unavailable but spot price succeeded. Null otherwise.
        instrument:
          type: object
          description: Honest labelling of the instrument, its trading mode, and data source.
          properties:
            kind:
              type: string
              enum: [CRYPTO, METAL]
            tradingMode:
              type: string
              enum: [ACTIVE, MONITORING, PAPER_UNVALIDATED]
            statusLabel:
              type: string
            currency:
              type: string
              enum: [GBP, USD]
            priceType:
              type: string
              enum: [SPOT, FUTURES]
            dataSource:
              type: string
          required: [kind, tradingMode, statusLabel, currency, priceType, dataSource]
      required: [coin, market, signal, oneHourTrend, fourHourTrend, indicators, position, metrics, risk, recentTrades, botStatus, message, instrument, opportunity]

    DirectionalEvaluation:
      type: object
      description: >
        Independent LONG and SHORT setup evaluation (all six assets).
        Crypto scores are weighted (max 8, gate >= 6); metals are raw condition
        counts (max 6; GOLD gate 5, SILVER gate 6). A direction qualifies when
        its score reaches its own threshold. PAPER TRADING ONLY.
      properties:
        longScore:
          type: number
        shortScore:
          type: number
        threshold:
          type: number
        shortThreshold:
          type: ["number", "null"]
        maxScore:
          type: ["number", "null"]
        decision:
          type: string
          enum: [LONG, SHORT, NO_TRADE]
        reason:
          oneOf:
            - type: string
            - type: "null"
        longConditions:
          type: array
          items:
            $ref: "#/components/schemas/StrategyCondition"
        shortConditions:
          type: array
          items:
            $ref: "#/components/schemas/StrategyCondition"
      required: [longScore, shortScore, threshold, decision, reason, longConditions, shortConditions]

    Opportunity:
      type: object
      description: Weighted opportunity score, market mode and entry status (paper mode).
      properties:
        score:
          type: number
        maxScore:
          type: number
        mode:
          type: string
          enum: [TREND, RANGE, DANGER]
        entryStatus:
          type: string
          enum: [READY, WAIT, BLOCKED]
        reason:
          oneOf:
            - type: string
            - type: "null"
        nextEligible:
          oneOf:
            - type: string
            - type: "null"
        lastTradeAt:
          oneOf:
            - type: string
            - type: "null"
      required: [score, maxScore, mode, entryStatus, reason, nextEligible, lastTradeAt]

    # ── Multi-coin state ────────────────────────────────────────────────────
    MultiCoinState:
      type: object
      properties:
        BTC:
          $ref: "#/components/schemas/PaperTraderState"
        ETH:
          $ref: "#/components/schemas/PaperTraderState"
        SOL:
          $ref: "#/components/schemas/PaperTraderState"
        XRP:
          $ref: "#/components/schemas/PaperTraderState"
        GOLD:
          $ref: "#/components/schemas/PaperTraderState"
        SILVER:
          $ref: "#/components/schemas/PaperTraderState"
      required: [BTC, ETH, SOL, XRP, GOLD, SILVER]

    # ── Engine (server-side scan scheduler) ─────────────────────────────────
    EngineStatus:
      type: object
      description: Status of the background strategy-scan scheduler in the API server.
      properties:
        status:
          type: string
          enum: [RUNNING, ERROR, STARTING]
        lastScanAt:
          type: ["string", "null"]
        nextScanAt:
          type: ["string", "null"]
        intervalSeconds:
          type: number
        lastError:
          type: ["string", "null"]
        scansCompleted:
          type: number
      required: [status, lastScanAt, nextScanAt, intervalSeconds, lastError, scansCompleted]

    ExecutionDiagnostics:
      type: object
      description: Explicit entry-eligibility report — every active blocker, named.
      properties:
        eligible:
          type: boolean
        blockers:
          type: array
          items:
            type: string
        lastCompletedCandleAt:
          type: ["string", "null"]
        checkedAt:
          type: ["string", "null"]
      required: [eligible, blockers]

    # ── Portfolio ───────────────────────────────────────────────────────────
    PortfolioSummary:
      type: object
      properties:
        openPositions:
          type: ["number", "null"]
        totalInstruments:
          type: ["number", "null"]
        totalOpenRisk:
          type: ["number", "null"]
          description: Sum of riskAmount across all open paper positions (£/$ aggregated 1:1).
        openRiskPercent:
          type: ["number", "null"]
        riskCeilingPercent:
          type: ["number", "null"]
        totalStarting:
          type: number
        totalBalance:
          type: number
        totalPnl:
          type: number
        totalRoi:
          type: number
        totalTrades:
          type: number
        totalWins:
          type: number
        totalLosses:
          type: number
        overallWinRate:
          type: number
        coins:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PaperTraderMetrics"
      required: [totalStarting, totalBalance, totalPnl, totalRoi, totalTrades, totalWins, totalLosses, overallWinRate, coins]

    # ── Activity log ────────────────────────────────────────────────────────
    ActivityEvent:
      type: object
      properties:
        id:
          type: number
        coin:
          type: string
        event:
          type: string
        message:
          type: string
        ts:
          type: string
      required: [id, coin, event, message, ts]

    # ── Input schemas ───────────────────────────────────────────────────────
    ResetPaperTraderInput:
      type: object
      properties:
        startingBalance:
          type: number
          minimum: 0