# lib/api-zod/src/generated/types/engineStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EngineStatusStatus } from './engineStatusStatus';

/**
 * Status of the background strategy-scan scheduler in the API server.
 */
export interface EngineStatus {
  status: EngineStatusStatus;
  /** @nullable */
  lastScanAt: string | null;
  /** @nullable */
  nextScanAt: string | null;
  intervalSeconds: number;
  /** @nullable */
  lastError: string | null;
  scansCompleted: number;
}