# lib/api-zod/src/generated/types/executionDiagnostics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Explicit entry-eligibility report — every active blocker, named.
 */
export interface ExecutionDiagnostics {
  eligible: boolean;
  blockers: string[];
  /** @nullable */
  lastCompletedCandleAt?: string | null;
  /** @nullable */
  checkedAt?: string | null;
}