# lib/api-zod/src/generated/types/paperTraderState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DirectionalEvaluation } from './directionalEvaluation';
import type { ExecutionDiagnostics } from './executionDiagnostics';
import type { IndicatorSnapshot } from './indicatorSnapshot';
import type { OpenPosition } from './openPosition';
import type { Opportunity } from './opportunity';
import type { PaperTrade } from './paperTrade';
import type { PaperTraderMetrics } from './paperTraderMetrics';
import type { PaperTraderStateBotStatus } from './paperTraderStateBotStatus';
import type { PaperTraderStateFourHourTrend } from './paperTraderStateFourHourTrend';
import type { PaperTraderStateInstrument } from './paperTraderStateInstrument';
import type { PaperTraderStateMarket } from './paperTraderStateMarket';
import type { PaperTraderStateOneHourTrend } from './paperTraderStateOneHourTrend';
import type { PaperTraderStateRisk } from './paperTraderStateRisk';
import type { PaperTraderStateSignal } from './paperTraderStateSignal';
import type { ProposedTrade } from './proposedTrade';
import type { StrategyConditions } from './strategyConditions';

export interface PaperTraderState {
  coin: string;
  market: PaperTraderStateMarket;
  signal: PaperTraderStateSignal;
  oneHourTrend: PaperTraderStateOneHourTrend;
  fourHourTrend: PaperTraderStateFourHourTrend;
  indicators: IndicatorSnapshot;
  strategyConditions?: StrategyConditions | null;
  proposedTrade?: ProposedTrade | null;
  directional?: DirectionalEvaluation | null;
  executionDiagnostics?: ExecutionDiagnostics | null;
  opportunity: Opportunity;
  position: OpenPosition | null;
  metrics: PaperTraderMetrics;
  risk: PaperTraderStateRisk;
  recentTrades: PaperTrade[];
  botStatus: PaperTraderStateBotStatus;
  message: string;
  /** Set when scan candle data (Yahoo Finance) is unavailable but spot price succeeded. Null otherwise. */
  scanNote?: string | null;
  /** Honest labelling of the instrument, its trading mode, and data source. */
  instrument: PaperTraderStateInstrument;
}